requirements.txt
streamlit
streamlit-authenticator
PyYAML
mysql-connector-python
pandas 